Extend RedisConnectionPoolService
